Begin replacing settingsmanager.py with simpler configparser
